requirements.txt
python-telegram-bot==20.6
gspread
google-auth
pandas